requirements/requirements_ai_vtuber.txt
apscheduler
fuzzywuzzy==0.18.0
python-Levenshtein
pandas
py2neo
neo4j
ffmpeg-python
edge_tts
zhipuai
langchain
langchain_community
sentence_transformers
chromadb
matplotlib
streamlit
pyaudio
faster_whisper
flask
flask_apscheduler
flask_cors
streamlit_webrtc
dashscope
openai
psutil
beautifulsoup4
lxml
yt-dlp




